feat: Refactor authentication routes and redirects
Update routes and ProtectedRoute component to improve user authentication flow.
If a user is already logged in, they are now redirected from the login page to the dashboard, and subsequently to the main landing page.
Additionally, users without the required role are redirected to the main landing page instead of the dashboard, simplifying navigation.